+++ includes/zshrc-files/zshrc-linux-snippet.zsh
@@ -7,7 +7,7 @@
 ###
 
 ## General aliases.
-alias zls="eza"
+alias python="python3"
 hash xdg-open 2>/dev/null && alias open="xdg-open"
 
 ## Configuration related aliases.